chore: code cleanup - move helper functions
